requirements.txt
Flask
gunicorn
Werkzeug
Jinja2
MarkupSafe
itsdangerous
click
azure-storage-blob